pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nameping"
dynamic = ["version"]
description = "Check name availability across domains and business registries"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Giacomo Marciani", email = "giacomo.marciani@gmail.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.12"
dependencies = [
    "click~=8.3.1",
    "Jinja2~=3.1.6",
    "pydantic~=2.12.5",
    "PyYAML~=6.0.3",
    "requests~=2.32.5",
]

[project.optional-dependencies]
dev = [
    "autoflake~=2.3",
    "black~=26.1",
    "build~=1.4",
    "flake8~=7.3",
    "mypy~=1.19",
    "pre-commit~=4.5",
    "pytest~=9.0",
    "pytest-cov~=7.0",
    "tox~=4.34",
    "twine~=6.2",
    "types-PyYAML~=6.0",
    "types-requests~=2.32",
]
docs = [
    "sphinx~=9.1",
    "sphinx-click~=6.2",
    "sphinx-rtd-theme~=3.1",
    "sphinx-new-tab-link~=0.8",
    "autodoc_pydantic~=2.2",
]

[project.scripts]
nameping = "nameping.cli:main"

[project.urls]
Homepage = "https://github.com/gmarciani/nameping"
Repository = "https://github.com/gmarciani/nameping"
Issues = "https://github.com/gmarciani/nameping/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["nameping*"]

[tool.setuptools.package-data]
nameping = [
    "config/*.yaml",
]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.coverage.run]
omit = ["*/__init__.py"]
